feat: account abstract base class for cash and credit accounts
