+++ src/api/accounts/credit/create_credit_account.py
@@ -5,7 +5,7 @@ from src.api.common.methods import WalterAPIMethod
 from src.api.common.models import Response, Status, HTTPStatus
 from src.auth.authenticator import WalterAuthenticator
 from src.aws.cloudwatch.client import WalterCloudWatchClient
-from src.database.credit_accounts.models import CreditAccount
+from src.database.accounts.credit.models import CreditAccount
 from src.database.client import WalterDB
 from src.database.users.models import User
 from src.utils.log import Logger
