+++ src/api/transactions/add_transaction.py
@@ -13,7 +13,7 @@ from src.api.common.methods import WalterAPIMethod
 from src.api.common.models import Response, HTTPStatus, Status
 from src.auth.authenticator import WalterAuthenticator
 from src.aws.cloudwatch.client import WalterCloudWatchClient
-from src.database.cash_accounts.models import CashAccount
+from src.database.accounts.cash.models import CashAccount
 from src.database.client import WalterDB
 from src.database.transactions.models import Transaction
 from src.utils.log import Logger
