+++ src/database/accounts/cash/models.py
@@ -1,8 +1,8 @@
 import datetime as dt
 import uuid
-from dataclasses import dataclass
 from enum import Enum
 
+from src.database.accounts.models import Account
 from src.database.users.models import User
 
 
@@ -23,8 +23,7 @@ class CashAccountType(Enum):
         raise ValueError(f"Unexpected cash account type '{account_type_str}'!")
 
 
-@dataclass
-class CashAccount:
+class CashAccount(Account):
     """
     WalterDB: CashAccount Model
 
@@ -36,10 +35,6 @@ class CashAccount:
     any expenses.
     """
 
-    USER_ID_KEY_FORMAT = "USER#{user_id}"
-    ACCOUNT_ID_KEY_FORMAT = "ACCOUNT#{account_id}"
-    DEFAULT_LOGO_URL = "https://walterai-public-media-dev.s3.us-east-1.amazonaws.com/cash-accounts/default/logo.svg"
-
     user_id: str  # hash key
     account_id: str  # sort key
     bank_name: str
@@ -49,65 +44,44 @@ class CashAccount:
     balance: float
     created_at: dt.datetime
     updated_at: dt.datetime
-    logo_url: str = DEFAULT_LOGO_URL
+
+    def __init__(
+        self,
+        user_id: str,
+        account_id: str,
+        bank_name: str,
+        account_name: str,
+        account_last_four_numbers: str,
+        balance: float,
+        created_at: dt.datetime,
+        updated_at: dt.datetime,
+        account_type: CashAccountType,
+        logo_url: str = Account.DEFAULT_LOGO_URL,
+    ) -> None:
+        super().__init__(
+            user_id,
+            account_id,
+            bank_name,
+            account_name,
+            account_last_four_numbers,
+            balance,
+            created_at,
+            updated_at,
+            logo_url,
+        )
+        self.account_type = account_type
 
     def to_dict(self) -> dict:
         return {
-            "account_id": self.account_id.split("#")[1],
-            "bank_name": self.bank_name,
-            "account_name": self.account_name,
+            **self._get_common_attributes_dict(),
             "account_type": self.account_type.value,
-            "account_last_four_numbers": self.account_last_four_numbers,
-            "balance": self.balance,
-            "created_at": self.created_at.isoformat(),
-            "updated_at": self.updated_at.isoformat(),
-            "logo_url": self.logo_url,
         }
 
     def to_ddb_item(self) -> dict:
         return {
-            "user_id": {
-                "S": self.user_id,
-            },
-            "account_id": {
-                "S": self.account_id,
-            },
-            "bank_name": {
-                "S": self.bank_name,
-            },
-            "account_name": {
-                "S": self.account_name,
-            },
-            "account_type": {
-                "S": self.account_type.value,
-            },
-            "account_last_four_numbers": {
-                "S": self.account_last_four_numbers,
-            },
-            "balance": {
-                "N": str(self.balance),
-            },
-            "created_at": {
-                "S": self.created_at.isoformat(),
-            },
-            "updated_at": {
-                "S": self.updated_at.isoformat(),
-            },
-            "logo_url": {
-                "S": self.logo_url,
-            },
+            **self._get_common_attributes_ddb_item(),
+            "account_type": {"S": self.account_type.value},
         }
-
-    def get_account_id(self) -> str:
-        return self.account_id.split("#")[1]
-
-    @staticmethod
-    def get_user_id_key(user_id: str) -> str:
-        return CashAccount.USER_ID_KEY_FORMAT.format(user_id=user_id)
-
-    @staticmethod
-    def get_account_id_key(account_id: str) -> str:
-        return CashAccount.ACCOUNT_ID_KEY_FORMAT.format(account_id=account_id)
 
     @classmethod
     def create_new_account(
