+++ src/database/client.py
@@ -4,10 +4,12 @@ from typing import Dict, List, Optional
 
 from src.auth.authenticator import WalterAuthenticator
 from src.aws.dynamodb.client import WalterDDBClient
-from src.database.cash_accounts.models import CashAccount
-from src.database.cash_accounts.table import CashAccountsTable
-from src.database.credit_accounts.models import CreditAccount
-from src.database.credit_accounts.table import CreditAccountsTable
+from src.database.accounts.cash.models import CashAccount
+from src.database.accounts.cash.table import CashAccountsTable
+from src.database.accounts.credit.models import CreditAccount
+from src.database.accounts.credit.table import CreditAccountsTable
+
+from src.database.accounts.models import Account
 from src.database.models import AccountTransaction
 from src.database.plaid_items.model import PlaidItem
 from src.database.plaid_items.table import PlaidItemsTable
@@ -229,6 +231,13 @@ class WalterDB:
             transaction_id: The unique ID of the transaction.
         """
         self.transactions_table.delete_transaction(user_id, date, transaction_id)
+
+    ############
+    # ACCOUNTS #
+    ############
+
+    def get_accounts(self, user_id: str) -> List[Account]:
+        pass
 
     #################
     # CASH ACCOUNTS #
