+++ walter.py
@@ -1,6 +1,6 @@
-from src.api.accounts.credit_accounts import CreateCreditAccount
-from src.api.accounts.credit_accounts import DeleteCreditAccount
-from src.api.accounts.credit_accounts import GetCreditAccounts
+from src.api.accounts.credit.create_credit_account import CreateCreditAccount
+from src.api.accounts.credit.delete_credit_account import DeleteCreditAccount
+from src.api.accounts.credit.get_credit_accounts import GetCreditAccounts
 from src.api.plaid.create_link_token import CreateLinkToken
 from src.api.plaid.exchange_public_token import ExchangePublicToken
 from src.api.plaid.refresh_transactions import RefreshTransactions
